Changed Produce Input to Text Field
